Updated data mapping
