+++ core/src/main/kotlin/com/tsobu/ona/core/dto/json/valdto/VaLPpTreatDto.kt
@@ -1,4 +1,4 @@
-package com.tsobu.ona.core.dto.json.`val`
+package com.tsobu.ona.core.dto.json.valdto
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
