+++ core/src/main/kotlin/com/tsobu/ona/core/dto/json/valdto/ValSphsOgDto.kt
@@ -1,4 +1,4 @@
-package com.tsobu.ona.core.dto.json.`val`
+package com.tsobu.ona.core.dto.json.valdto
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder("SubmissionDate", "uuid", "start", "today", "deviceid", "subscriberid", "email", "username", "simserial", "phonenumber", "banner", "intro", "intro2", "VAL", "EAID", "HHID", "countrySelect", "country", "currency", "conversion", "geopoint-Latitude", "geopoint-Longitude", "geopoint-Altitude", "geopoint-Accuracy", "lat", "lon", "plantingDate", "season", "plantingWeek0", "harvestDate", "harvestWeek0", "harvestWeekm1", "harvestWeekm2", "harvestWeekm3", "harvestWeekm4", "harvestWeekp1", "harvestWeekp2", "harvestWeekp3", "harvestWeekp4", "harvestAge0", "harvestAgem1", "harvestAgem2", "harvestAgem3", "harvestAgem4", "harvestAgep1", "harvestAgep2", "harvestAgep3", "harvestAgep4", "check_id", "check", "variety", "plantingDensity", "betweenRow", "withinRow", "areaBasis", "plotL", "plotW", "plotSize", "yieldLevel", "priceStarch", "starchFactory", "SC1", "SC2", "SC3", "SC4", "SC5", "SC6", "SC7", "SC8", "SC9", "SC10", "SC11", "SC12", "priceSC1", "priceSC2", "priceSC3", "priceSC4", "priceSC5", "priceSC6", "priceSC7", "priceSC8", "priceSC9", "priceSC10", "priceSC11", "priceSC12", "product", "transformRatio", "unit", "unitNameSelect", "unitNameOther", "unitWeightOther", "unitName", "unitLocalWeight", "unitWeight", "priceNote", "priceRm4", "priceRm2", "priceR0", "priceRp2", "priceRp4", "price0", "pricem1", "pricem2", "pricem3", "pricem4", "pricep1", "pricep2", "pricep3", "pricep4", "noRecommendationNote", "lookupp0h0_id", "lookupp0hm1_id", "lookupp0hm2_id", "lookupp0hm3_id", "lookupp0hm4_id", "lookupp0hp1_id", "lookupp0hp2_id", "lookupp0hp3_id", "lookupp0hp4_id", "WYp0h0", "WYp0hm1", "WYp0hm2", "WYp0hm3", "WYp0hm4", "WYp0hp1", "WYp0hp2", "WYp0hp3", "WYp0hp4", "CYp0h0", "CYp0hm1", "CYp0hm2", "CYp0hm3", "CYp0hm4", "CYp0hp1", "CYp0hp2", "CYp0hp3", "CYp0hp4", "SC0", "SCm1", "SCm2", "SCm3", "SCm4", "SCp1", "SCp2", "SCp3", "SCp4", "priceSC0", "priceSCm1", "priceSCm2", "priceSCm3", "priceSCm4", "priceSCp1", "priceSCp2", "priceSCp3", "priceSCp4", "Yp0h0", "Yp0h0_AB", "Yp0hm1", "Yp0hm2", "Yp0hm3", "Yp0hm4", "Yp0hp1", "Yp0hp2", "Yp0hp3", "Yp0hp4", "GRp0h0", "GRp0h0_AB", "GRp0h0_ABcurr", "GRp0hm1", "GRp0hm2", "GRp0hm3", "GRp0hm4", "GRp0hp1", "GRp0hp2", "GRp0hp3", "GRp0hp4", "Pp0h0", "Pp0hm1", "Pp0hm2", "Pp0hm3", "Pp0hm4", "Pp0hp1", "Pp0hp2", "Pp0hp3", "Pp0hp4", "GPp0h0", "GPp0h0curr", "GPp0hm1", "GPp0hm2", "GPp0hm3", "GPp0hm4", "GPp0hp1", "GPp0hp2", "GPp0hp3", "GPp0hp4", "GRmax", "GRmax_AB", "GRdif", "GRdif_AB", "GRdif_ABcurr", "GPmax", "GPdif", "GPdifcurr", "optHarvest", "optHarvestText", "Yh0p0Note", "optHarvestNote", "difHarvestNote", "GRdifNote", "GRmaxNote", "sensible", "validationNote", "validationNote2", "validationNote3", "call", "confirmVAL", "end", "instanceID", "KEY")
-class ValSphsKwDto {
+class ValSphsOgDto {
     @JsonProperty("SubmissionDate")
     var submissionDate: String? = null
 
@@ -68,16 +68,16 @@ class ValSphsKwDto {
     var conversion: String? = null
 
     @JsonProperty("geopoint-Latitude")
-    var geopointLatitude: String? = null
+    var geoPointLatitude: String? = null
 
     @JsonProperty("geopoint-Longitude")
-    var geopointLongitude: String? = null
+    var geoPointLongitude: String? = null
 
     @JsonProperty("geopoint-Altitude")
-    var geopointAltitude: String? = null
+    var geoPointAltitude: String? = null
 
     @JsonProperty("geopoint-Accuracy")
-    var geopointAccuracy: String? = null
+    var geoPointAccuracy: String? = null
 
     @JsonProperty("lat")
     var lat: String? = null
@@ -209,10 +209,10 @@ class ValSphsKwDto {
     var sc6: String? = null
 
     @JsonProperty("SC7")
-    var sc7: String? = null
+    var sc7: Any? = null
 
     @JsonProperty("SC8")
-    var sc8: String? = null
+    var sc8: Any? = null
 
     @JsonProperty("SC9")
     var sc9: String? = null
@@ -245,10 +245,10 @@ class ValSphsKwDto {
     var priceSc6: String? = null
 
     @JsonProperty("priceSC7")
-    var priceSc7: String? = null
+    var priceSc7: Any? = null
 
     @JsonProperty("priceSC8")
-    var priceSc8: String? = null
+    var priceSc8: Any? = null
 
     @JsonProperty("priceSC9")
     var priceSc9: String? = null
@@ -266,7 +266,7 @@ class ValSphsKwDto {
     var product: String? = null
 
     @JsonProperty("transformRatio")
-    var transformRatio: String? = null
+    var transformRatio: Any? = null
 
     @JsonProperty("unit")
     var unit: String? = null
@@ -287,52 +287,52 @@ class ValSphsKwDto {
     var unitLocalWeight: Any? = null
 
     @JsonProperty("unitWeight")
-    var unitWeight: String? = null
+    var unitWeight: Any? = null
 
     @JsonProperty("priceNote")
     var priceNote: String? = null
 
     @JsonProperty("priceRm4")
-    var priceRm4: String? = null
+    var priceRm4: Any? = null
 
     @JsonProperty("priceRm2")
-    var priceRm2: String? = null
+    var priceRm2: Any? = null
 
     @JsonProperty("priceR0")
-    var priceR0: String? = null
+    var priceR0: Any? = null
 
     @JsonProperty("priceRp2")
-    var priceRp2: String? = null
+    var priceRp2: Any? = null
 
     @JsonProperty("priceRp4")
-    var priceRp4: String? = null
+    var priceRp4: Any? = null
 
     @JsonProperty("price0")
-    var price0: String? = null
+    var price0: Any? = null
 
     @JsonProperty("pricem1")
-    var priceM1: String? = null
+    var priceM1: Any? = null
 
     @JsonProperty("pricem2")
-    var priceM2: String? = null
+    var priceM2: Any? = null
 
     @JsonProperty("pricem3")
-    var priceM3: String? = null
+    var priceM3: Any? = null
 
     @JsonProperty("pricem4")
-    var priceM4: String? = null
+    var priceM4: Any? = null
 
     @JsonProperty("pricep1")
-    var priceP1: String? = null
+    var priceP1: Any? = null
 
     @JsonProperty("pricep2")
-    var priceP2: String? = null
+    var priceP2: Any? = null
 
     @JsonProperty("pricep3")
-    var priceP3: String? = null
+    var priceP3: Any? = null
 
     @JsonProperty("pricep4")
-    var priceP4: String? = null
+    var priceP4: Any? = null
 
     @JsonProperty("noRecommendationNote")
     var noRecommendationNote: String? = null
@@ -383,7 +383,7 @@ class ValSphsKwDto {
     var wyP0hp1: String? = null
 
     @JsonProperty("WYp0hp2")
-    var wyP0hp2: String? = null
+    var wyP0hp2: Any? = null
 
     @JsonProperty("WYp0hp3")
     var wyP0hp3: Any? = null
@@ -410,7 +410,7 @@ class ValSphsKwDto {
     var cyP0hp1: String? = null
 
     @JsonProperty("CYp0hp2")
-    var cyP0hp2: String? = null
+    var cyP0hp2: Any? = null
 
     @JsonProperty("CYp0hp3")
     var cyP0hp3: Any? = null
@@ -476,7 +476,7 @@ class ValSphsKwDto {
     var yp0h0: String? = null
 
     @JsonProperty("Yp0h0_AB")
-    var yp0h0AB: String? = null
+    var yp0h0Ab: String? = null
 
     @JsonProperty("Yp0hm1")
     var yp0hm1: String? = null
@@ -494,7 +494,7 @@ class ValSphsKwDto {
     var yp0hp1: String? = null
 
     @JsonProperty("Yp0hp2")
-    var yp0hp2: String? = null
+    var yp0hp2: Any? = null
 
     @JsonProperty("Yp0hp3")
     var yp0hp3: Any? = null
@@ -506,7 +506,7 @@ class ValSphsKwDto {
     var grP0h0: String? = null
 
     @JsonProperty("GRp0h0_AB")
-    var grP0h0AB: String? = null
+    var grP0h0Ab: String? = null
 
     @JsonProperty("GRp0h0_ABcurr")
     var grP0h0ABCurr: String? = null
@@ -527,7 +527,7 @@ class ValSphsKwDto {
     var grP0hp1: String? = null
 
     @JsonProperty("GRp0hp2")
-    var grP0hp2: String? = null
+    var grP0hp2: Any? = null
 
     @JsonProperty("GRp0hp3")
     var grP0hp3: Any? = null
@@ -554,7 +554,7 @@ class ValSphsKwDto {
     var pp0hp1: String? = null
 
     @JsonProperty("Pp0hp2")
-    var pp0hp2: String? = null
+    var pp0hp2: Any? = null
 
     @JsonProperty("Pp0hp3")
     var pp0hp3: Any? = null
@@ -584,7 +584,7 @@ class ValSphsKwDto {
     var gpP0hp1: String? = null
 
     @JsonProperty("GPp0hp2")
-    var gpP0hp2: String? = null
+    var gpP0hp2: Any? = null
 
     @JsonProperty("GPp0hp3")
     var gpP0hp3: Any? = null
@@ -596,16 +596,16 @@ class ValSphsKwDto {
     var grMax: String? = null
 
     @JsonProperty("GRmax_AB")
-    var grMaxAB: String? = null
+    var grMaxAb: String? = null
 
     @JsonProperty("GRdif")
     var grDif: String? = null
 
     @JsonProperty("GRdif_AB")
-    var grDifAB: String? = null
+    var grDifAb: String? = null
 
     @JsonProperty("GRdif_ABcurr")
-    var grDifABCurr: String? = null
+    var grDifAbCurr: String? = null
 
     @JsonProperty("GPmax")
     var gpMax: String? = null
