+++ core/src/main/kotlin/com/tsobu/ona/core/service/valform/ValPpService.kt
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.MapperFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.tsobu.ona.core.config.AppConfig
-import com.tsobu.ona.core.dto.json.`val`.ValPpDto
-import com.tsobu.ona.core.dto.json.`val`.ValPpPwDto
+import com.tsobu.ona.core.dto.json.valdto.ValPpDto
+import com.tsobu.ona.core.dto.json.valdto.ValPpPwDto
 import com.tsobu.ona.core.utils.MyUtils
 import com.tsobu.ona.core.utils.CsvUtility
 import com.tsobu.ona.database.entities.`val`.ValPpEntity
