+++ core/src/main/kotlin/com/tsobu/ona/core/service/valsphs/ValSphsOyService.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.MapperFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.tsobu.ona.core.config.AppConfig
-import com.tsobu.ona.core.dto.json.`val`.ValSphsOyDto
+import com.tsobu.ona.core.dto.json.valdto.ValSphsOyDto
 import com.tsobu.ona.core.utils.CsvUtility
 import com.tsobu.ona.core.utils.MyUtils
 import com.tsobu.ona.database.entities.`val`.ValSphsOyEntity
